requirements.txt
aiofiles==23.2.1
asyncpg==0.29.0
fastapi==0.110.1
python-dotenv==0.19.2
SQLAlchemy==2.0.29
uvicorn==0.15.0
alembic==1.13.1
python-multipart==0.0.9
